Knex seed:make and migrate:make now create the files in /src, rather than /build (which then required moving)
Use babel-node to run knex, so it can read the knexfile that is written in es6 (instead of having to transpile it first). This means the new seed and migrate files are created in the right place (in /src/db). We only create seed and migration files on local env, so don't need to worry about doing this in the same way as on prod.